lru_crawler_tocrawl to limit inspected per run
Since we start from the tail, and some slabs could be many millions of items,
it doesn't always make sense to walk the entire thing looking for data to use.

This could be a lot smarter: say specify a percentage, and it'll only crawl a
percentage of used_chunks for that slab for instance.